feat: reworked determineBaseExercise method
Reworked determineBaseExercise method of ExerciseTypeService. In an
earlier version this method used a switch statement to determine the
baseExercise. This method has been rewritten and uses a complex if
statement to match certain strings to the exercise name to determine the
baseExercise.
This method of the service class has been implemented in the CsvService.